+++ clang/heap/read.c
@@ -1,4 +1,4 @@
-fn Term heap_read(u32 loc) {
+fn Term heap_read(u64 loc) {
   Term term = __atomic_load_n(&HEAP[loc], __ATOMIC_RELAXED);
   if (__builtin_expect(term != 0, 1)) {
     return term;
