+++ clang/heap/set.c
@@ -1,3 +1,3 @@
-fn void heap_set(u32 loc, Term val) {
+fn void heap_set(u64 loc, Term val) {
   __atomic_store_n(&HEAP[loc], val, __ATOMIC_RELAXED);
 }
