+++ clang/heap/subst_var.c
@@ -1,3 +1,3 @@
-fn void heap_subst_var(u32 loc, Term val) {
+fn void heap_subst_var(u64 loc, Term val) {
   heap_set(loc, term_sub_set(val, 1));
 }
