+++ clang/parse/term/var.c
@@ -18,7 +18,7 @@ fn Term parse_term_var(PState *s, u32 depth) {
     u32 offset = (side == 1) ? 1 : 0;
     return term_new(0, BJV, 0, (u32)bind->lvl + offset);
   }
-  u32 val = (u32)bind->lvl;
+  u64 val = (u32)bind->lvl;
   u32 lab = bind->lab;
   u8  tag = (side == 0) ? BJ0 : (side == 1) ? BJ1 : BJV;
   return term_new(0, tag, lab, val);
