+++ clang/term/clone.c
@@ -1,4 +1,4 @@
-fn Copy term_clone_at(u32 loc, u32 lab) {
+fn Copy term_clone_at(u64 loc, u32 lab) {
   return (Copy){ term_new_dp0(lab, loc), term_new_dp1(lab, loc) };
 }
