+++ clang/term/new/dry.c
@@ -1,4 +1,4 @@
-fn Term term_new_dry_at(u32 loc, Term fun, Term arg) {
+fn Term term_new_dry_at(u64 loc, Term fun, Term arg) {
   return term_new_at(loc, DRY, 0, 2, (Term[]){fun, arg});
 }
