+++ clang/term/new/lam.c
@@ -1,4 +1,4 @@
-fn Term term_new_lam_at(u32 loc, Term bod) {
+ fn Term term_new_lam_at(u64 loc, Term bod) {
   return term_new_at(loc, LAM, 0, 1, (Term[]){bod});
 }
