+++ clang/term/new/uns.c
@@ -1,7 +1,7 @@
 // UNS: ! ${f, v}; body - unscoped binding
 // fields = [body]
 fn Term term_new_uns(Term bod) {
-  u32 loc = heap_alloc(1);
+  u64 loc = heap_alloc(1);
   heap_set(loc, bod);
   return term_new(0, UNS, 0, loc);
 }
