+++ clang/term/new/use.c
@@ -1,6 +1,6 @@
 // USE: λ{f} - reduce arg and apply
 // fields = [f]
-fn Term term_new_use_at(u32 loc, Term f) {
+fn Term term_new_use_at(u64 loc, Term f) {
   heap_set(loc, f);
   return term_new(0, USE, 0, loc);
 }
