+++ clang/term/new/var.c
@@ -1,3 +1,3 @@
-fn Term term_new_var(u32 loc) {
+fn Term term_new_var(u64 loc) {
   return term_new(0, VAR, 0, loc);
 }
