+++ clang/term/val.c
@@ -1,3 +1,3 @@
-fn u32 term_val(Term t) {
+fn u64 term_val(Term t) {
   return (t >> VAL_SHIFT) & VAL_MASK;
 }
