+++ clang/wnf/and_inc.c
@@ -3,7 +3,7 @@
 // ↑(a & b)
 fn Term wnf_and_inc(Term inc, Term b) {
   ITRS_INC("AND-INC");
-  u32  inc_loc = term_val(inc);
+  u64  inc_loc = term_val(inc);
   Term a       = heap_read(inc_loc);
   Term and_tm  = term_new_and(a, b);
   heap_set(inc_loc, and_tm);
