+++ clang/wnf/and_sup.c
@@ -5,7 +5,7 @@
 fn Term wnf_and_sup(Term sup, Term b) {
   ITRS_INC("AND-SUP");
   u32  lab = term_ext(sup);
-  u32  loc = term_val(sup);
+  u64  loc = term_val(sup);
   Term a0  = heap_read(loc + 0);
   Term a1  = heap_read(loc + 1);
   u64  dup_loc = heap_alloc(1);
