+++ clang/wnf/app_lam.c
@@ -4,7 +4,7 @@
 // f
 fn Term wnf_app_lam(Term lam, Term arg) {
   ITRS_INC("APP-LAM");
-  u32  loc     = term_val(lam);
+  u64  loc     = term_val(lam);
   u32  lam_ext = term_ext(lam);
   Term body    = heap_read(loc);
   if (lam_ext & LAM_ERA_MASK) {
