+++ clang/wnf/app_mat_ctr.c
@@ -8,7 +8,7 @@
 fn Term wnf_app_mat_ctr(Term mat, Term ctr) {
   u32 mat_ext = term_ext(mat);
   u32 ctr_ext = term_ext(ctr);
-  u32 mat_loc = term_val(mat);
+  u64 mat_loc = term_val(mat);
   if (mat_ext == ctr_ext) {
     ITRS_INC("APP-MAT-CTR-MAT");
     u32 ari = term_tag(ctr) - C00;
@@ -16,7 +16,7 @@ fn Term wnf_app_mat_ctr(Term mat, Term ctr) {
     if (ari == 0) {
       return res;
     }
-    u32 ctr_loc = term_val(ctr);
+    u64 ctr_loc = term_val(ctr);
     // Reuse MAT node storage for the first APP in the chain.
     res = term_new_app_at(mat_loc, res, heap_read(ctr_loc + 0));
     if (ari == 1) {
@@ -24,7 +24,7 @@ fn Term wnf_app_mat_ctr(Term mat, Term ctr) {
     }
     u64 apps = heap_alloc(2 * (u64)(ari - 1));
     for (u32 i = 1; i < ari; i++) {
-      res = term_new_app_at((u32)(apps + 2 * (u64)(i - 1)), res, heap_read(ctr_loc + i));
+      res = term_new_app_at(apps + 2 * (u64)(i - 1), res, heap_read(ctr_loc + i));
     }
     return res;
   } else {
