+++ clang/wnf/app_mat_num.c
@@ -6,9 +6,9 @@
 // --------------- APP-MAT-NUM-MIS
 // (m #b)
 fn Term wnf_app_mat_num(Term mat, Term num) {
-  u32 mat_loc = term_val(mat);
+  u64 mat_loc = term_val(mat);
   u32 mat_ext = term_ext(mat);
-  u32 num_val = term_val(num);
+  u64 num_val = term_val(num);
   if (mat_ext == num_val) {
     ITRS_INC("APP-MAT-NUM-MAT");
     return heap_read(mat_loc + 0);
