+++ clang/wnf/eql_inc.c
@@ -3,7 +3,7 @@
 // ↑(a === b)
 fn Term wnf_eql_inc_l(Term inc, Term b) {
   ITRS_INC("EQL-INC-L");
-  u32  inc_loc = term_val(inc);
+  u64  inc_loc = term_val(inc);
   Term a       = heap_read(inc_loc);
   Term eql     = term_new_eql(a, b);
   heap_set(inc_loc, eql);
@@ -15,7 +15,7 @@ fn Term wnf_eql_inc_l(Term inc, Term b) {
 // ↑(a === b)
 fn Term wnf_eql_inc_r(Term a, Term inc) {
   ITRS_INC("EQL-INC-R");
-  u32  inc_loc = term_val(inc);
+  u64  inc_loc = term_val(inc);
   Term b       = heap_read(inc_loc);
   Term eql     = term_new_eql(a, b);
   heap_set(inc_loc, eql);
