+++ clang/wnf/eql_lam.c
@@ -6,8 +6,8 @@
 // af === bf
 fn Term wnf_eql_lam(Term a, Term b) {
   ITRS_INC("EQL-LAM");
-  u32  a_loc = term_val(a);
-  u32  b_loc = term_val(b);
+  u64  a_loc = term_val(a);
+  u64  b_loc = term_val(b);
   Term af    = heap_read(a_loc);
   Term bf    = heap_read(b_loc);
   // Generate fresh name for substitution
